chore(tests): Move tests from src directory
