+++ tests/fixtures/entities/Base.ts
