+++ vitest.config.ts
@@ -2,12 +2,11 @@ import {defineConfig} from "vitest/config"
 
 export default defineConfig({
   test: {
-    include: ["src/**/*.test.ts"],
+    include: ["tests/**/*.test.ts"],
     reporters: ["default", "junit"],
     outputFile: "./vitest-report.junit.xml",
     coverage: {
-      include: ["src"],
-      exclude: ["src/tests/**"]
+      include: ["src/**/*.ts"]
     }
   }
 })
